refactor: CarController is its own module now! works fr after ./gradlew clean :)
