+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/HomeController.java
@@ -4,9 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class
-
-HomeController {
+public class HomeController {
 
     @GetMapping("/")
     public String homePage() {
